chore: seperate file for VariablesNormalizer
