+++ v2/pkg/astnormalization/astnormalization.go
@@ -345,76 +345,6 @@ func (o *OperationNormalizer) NormalizeNamedOperation(operation, definition *ast
 	}
 }
 
-type VariablesNormalizer struct {
-	firstDetectUnused          *astvisitor.Walker
-	secondExtract              *astvisitor.Walker
-	thirdDeleteUnused          *astvisitor.Walker
-	fourthCoerce               *astvisitor.Walker
-	variablesExtractionVisitor *variablesExtractionVisitor
-}
-
-// NewVariablesNormalizer creates a new variable normalizer.
-// If withFieldArgMapping is true then the normalizer creates a
-// mapping of field arguments as a "side product" when NormalizeOperation is called.
-func NewVariablesNormalizer(withFieldArgMapping bool) *VariablesNormalizer {
-	// delete unused modifying variables refs,
-	// so it is safer to run it sequentially with the extraction
-	thirdDeleteUnused := astvisitor.NewWalkerWithID(8, "DeleteUnusedVariables")
-	del := deleteUnusedVariables(&thirdDeleteUnused)
-
-	// register variable usage detection on the first stage
-	// and pass usage information to the deletion visitor
-	// so it keeps variables that are defined but not used at all
-	// ensuring that validation can still catch them
-	firstDetectUnused := astvisitor.NewWalkerWithID(8, "DetectVariableUsage")
-	detectVariableUsage(&firstDetectUnused, del)
-
-	secondExtract := astvisitor.NewWalkerWithID(8, "ExtractVariables")
-	variablesExtractionVisitor := extractVariables(&secondExtract, withFieldArgMapping)
-	extractVariablesDefaultValue(&secondExtract)
-
-	fourthCoerce := astvisitor.NewWalkerWithID(0, "VariablesCoercion")
-	inputCoercionForList(&fourthCoerce)
-
-	return &VariablesNormalizer{
-		firstDetectUnused:          &firstDetectUnused,
-		secondExtract:              &secondExtract,
-		thirdDeleteUnused:          &thirdDeleteUnused,
-		fourthCoerce:               &fourthCoerce,
-		variablesExtractionVisitor: variablesExtractionVisitor,
-	}
-}
-
-// NormalizeOperation processes GraphQL operation variables.
-// It detects and removes unused variables, extracts variables from inline values
-// and coerces variable values.
-// https://spec.graphql.org/September2025/#sec-Coercing-Variable-Values
-// It modifies the operation in place and
-// returns metadata including field argument mappings and upload paths.
-// Field argument mapping is done only when it is enabled in v,
-// else VariablesNormalizerResult.FieldArgumentMapping will be nil.
-// Any errors encountered during normalization are reported via the report parameter.
-func (v *VariablesNormalizer) NormalizeOperation(operation, definition *ast.Document, report *operationreport.Report) VariablesNormalizationResult {
-	v.firstDetectUnused.Walk(operation, definition, report)
-	if report.HasErrors() {
-		return VariablesNormalizationResult{}
-	}
-	v.secondExtract.Walk(operation, definition, report)
-	if report.HasErrors() {
-		return VariablesNormalizationResult{}
-	}
-	v.thirdDeleteUnused.Walk(operation, definition, report)
-	if report.HasErrors() {
-		return VariablesNormalizationResult{}
-	}
-	v.fourthCoerce.Walk(operation, definition, report)
-
-	return VariablesNormalizationResult{
-		UploadsMapping:       v.variablesExtractionVisitor.uploadsPath,
-		FieldArgumentMapping: v.variablesExtractionVisitor.fieldArgumentMapping.result,
-	}
-}
-
 type fragmentCycleVisitor struct {
 	*astvisitor.Walker
